Move InstrumentationLibraryInfo to common to be shared with metrics.
