+++ sdk/src/main/java/io/opentelemetry/sdk/trace/ReadableSpan.java
@@ -16,6 +16,7 @@
 
 package io.opentelemetry.sdk.trace;
 
+import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.trace.Span;
 import io.opentelemetry.trace.SpanContext;
