+++ sdk/src/main/java/io/opentelemetry/sdk/trace/SpanBuilderSdk.java
@@ -18,6 +18,7 @@ package io.opentelemetry.sdk.trace;
 
 import io.opentelemetry.internal.Utils;
 import io.opentelemetry.sdk.common.Clock;
+import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.internal.MonotonicClock;
 import io.opentelemetry.sdk.resources.Resource;
 import io.opentelemetry.sdk.trace.Sampler.Decision;
